List files

See merge request collint9/2021-ca400-collint9-coynemt2!28

+++ src/server/data/repo/file.go
@@ -5,7 +5,7 @@ import (
 	"gitlab.computing.dcu.ie/collint9/2021-ca400-collint9-coynemt2/src/server/data/schema"
 )
 
-func GetFilenameByFileId(fileId int64) (string, error) {
+func GetFilenameByFileId(fileId uint64) (string, error) {
 	var file schema.File
 	tx := data.DB.Begin()
 	if res := tx.First(&file, fileId); res.Error != nil {
@@ -14,7 +14,7 @@ func GetFilenameByFileId(fileId int64) (string, error) {
 	return file.Name, nil
 }
 
-func GetFileOwnerByFileId(fileId int64) (uint, error) {
+func GetFileOwnerByFileId(fileId uint64) (uint, error) {
 	var file schema.File
 	tx := data.DB.Begin()
 	if res := tx.First(&file, fileId); res.Error != nil {
@@ -23,7 +23,7 @@ func GetFileOwnerByFileId(fileId int64) (uint, error) {
 	return file.UserID, nil
 }
 
-func UpdateFileName(uploadRequestId uint, uniqueFilename string) {
+func UpdateFileName(uploadRequestId uint64, uniqueFilename string) {
 	var file schema.File
 	tx := data.DB.Begin()
 	tx.Model(&file).Where("id", uploadRequestId).Update("name", uniqueFilename)
